requirements.txt
discord.py
openai
asyncpg
python-dotenv
watchdog